+++ mediation/Vungle/source/plugin/Assets/GoogleMobileAds/Platforms/Android/Mediation/Vungle/VungleClient.cs
@@ -27,6 +27,9 @@ namespace GoogleMobileAds.Android.Mediation.Vungle
         private static VungleClient instance = new VungleClient();
         private VungleClient() {}
 
+        private const string VUNGLE_CLASS_NAME = "com.vungle.warren.Vungle";
+        private const string VUNGLE_CONSENT_ENUM_NAME = "com.vungle.warren.Vungle$Consent";
+
         public static VungleClient Instance
         {
             get
@@ -35,62 +38,98 @@ namespace GoogleMobileAds.Android.Mediation.Vungle
             }
         }
 
-        public void UpdateConsentStatus(VungleConsent consentStatus)
-        {
-            UpdateConsentStatus(consentStatus, "");
-        }
-
         public void UpdateConsentStatus(VungleConsent consentStatus,
                                         String consentMessageVersion)
         {
-            if (consentStatus == VungleConsent.UNKNOWN) {
-                MonoBehaviour.print("Cannot call 'VungleConsent.updateConsentStatus()' with unknown consent status.");
+            if (consentStatus == VungleConsent.UNKNOWN)
+            {
+                MonoBehaviour.print("[Vungle Plugin] Cannot call 'UpdateConsentStatus()' " +
+                        "with unknown consent status.");
                 return;
             }
-            if (consentMessageVersion == null) {
-                consentMessageVersion = "";
-            }
 
-            AndroidJavaObject vungleConsentObject;
-            AndroidJavaClass vungleConsentEnum = new AndroidJavaClass("com.vungle.warren.Vungle$Consent");
-            if (consentStatus == VungleConsent.ACCEPTED) {
-                vungleConsentObject = vungleConsentEnum.GetStatic<AndroidJavaObject>("OPTED_IN");
-            } else {
-                vungleConsentObject = vungleConsentEnum.GetStatic<AndroidJavaObject>("OPTED_OUT");
-            }
-
-            AndroidJavaClass vungleConsentClass = new AndroidJavaClass("com.vungle.mediation.VungleConsent");
-            vungleConsentClass.CallStatic("updateConsentStatus", vungleConsentObject, consentMessageVersion);
+            AndroidJavaObject vungleConsentStatusObject =
+                    GetConsentStatusFromVungleConsent(consentStatus);
+            AndroidJavaClass vungle = new AndroidJavaClass(VUNGLE_CLASS_NAME);
+            vungle.CallStatic("updateConsentStatus",
+                    vungleConsentStatusObject, consentMessageVersion);
         }
 
         public VungleConsent GetCurrentConsentStatus()
         {
-            AndroidJavaClass vungleConsentClass = new AndroidJavaClass("com.vungle.mediation.VungleConsent");
-            AndroidJavaClass vungleConsentEnum = new AndroidJavaClass("com.vungle.warren.Vungle$Consent");
-
-            AndroidJavaObject vungleConsent = vungleConsentClass.CallStatic<AndroidJavaObject>("getCurrentVungleConsent");
-
-            if (vungleConsent == null) {
-                return VungleConsent.UNKNOWN;
-            }
-
-            int vungleConsentValue = vungleConsent.Call<int> ("ordinal");
-            int optedInConsentValue = vungleConsentEnum.GetStatic<AndroidJavaObject>("OPTED_IN").Call<int>("ordinal");
-            int optedOutConsentValue = vungleConsentEnum.GetStatic<AndroidJavaObject>("OPTED_OUT").Call<int>("ordinal");
-
-            if (vungleConsentValue == optedInConsentValue) {
-                return VungleConsent.ACCEPTED;
-            } else if (vungleConsentValue == optedOutConsentValue) {
-                return VungleConsent.DENIED;
-            }
-
-            return VungleConsent.UNKNOWN;
+            AndroidJavaClass vungle = new AndroidJavaClass(VUNGLE_CLASS_NAME);
+            AndroidJavaObject consentStatus = vungle.CallStatic<AndroidJavaObject>("getConsentStatus");
+            return GetVungleConsentFromConsentStatus(consentStatus);
         }
 
         public String GetCurrentConsentMessageVersion()
         {
-            AndroidJavaClass vungleConsentClass = new AndroidJavaClass("com.vungle.mediation.VungleConsent");
-            return vungleConsentClass.CallStatic<String>("getCurrentVungleConsentMessageVersion");
+            AndroidJavaClass vungle = new AndroidJavaClass(VUNGLE_CLASS_NAME);
+            return vungle.CallStatic<String>("getConsentMessageVersion");
+        }
+
+        public void UpdateCCPAStatus(VungleConsent consentStatus)
+        {
+            if (consentStatus == VungleConsent.UNKNOWN)
+            {
+                MonoBehaviour.print("[Vungle Plugin] Cannot call 'UpdateCCPAStatus()' " +
+                        "with unknown consent status.");
+                return;
+            }
+
+            AndroidJavaObject vungleConsentStatusObject =
+                    GetConsentStatusFromVungleConsent(consentStatus);
+            AndroidJavaClass vungle = new AndroidJavaClass(VUNGLE_CLASS_NAME);
+            vungle.CallStatic("updateCCPAStatus", vungleConsentStatusObject);
+        }
+
+        public VungleConsent GetCCPAStatus()
+        {
+            AndroidJavaClass vungle = new AndroidJavaClass(VUNGLE_CLASS_NAME);
+            AndroidJavaObject consentStatus = vungle.CallStatic<AndroidJavaObject>("getCCPAStatus");
+            return GetVungleConsentFromConsentStatus(consentStatus);
+        }
+
+        private AndroidJavaObject GetConsentStatusFromVungleConsent(VungleConsent vungleConsent)
+        {
+            AndroidJavaClass vungleConsentEnum = new AndroidJavaClass(VUNGLE_CONSENT_ENUM_NAME);
+            switch (vungleConsent)
+            {
+                case VungleConsent.ACCEPTED:
+                    return vungleConsentEnum.GetStatic<AndroidJavaObject>("OPTED_IN");
+                case VungleConsent.DENIED:
+                    return vungleConsentEnum.GetStatic<AndroidJavaObject>("OPTED_OUT");
+                default:
+                    return null;
+            }
+        }
+
+        private VungleConsent GetVungleConsentFromConsentStatus(AndroidJavaObject consentStatus)
+        {
+            if (consentStatus == null)
+            {
+                return VungleConsent.UNKNOWN;
+            }
+
+            int vungleConsentValue = consentStatus.Call<int>("ordinal");
+
+            AndroidJavaClass vungleConsentEnum = new AndroidJavaClass(VUNGLE_CONSENT_ENUM_NAME);
+            int optedInConsentValue = vungleConsentEnum
+                    .GetStatic<AndroidJavaObject>("OPTED_IN")
+                    .Call<int>("ordinal");
+            int optedOutConsentValue = vungleConsentEnum
+                    .GetStatic<AndroidJavaObject>("OPTED_OUT")
+                    .Call<int>("ordinal");
+
+            if (vungleConsentValue == optedInConsentValue)
+            {
+                return VungleConsent.ACCEPTED;
+            }
+            else if (vungleConsentValue == optedOutConsentValue)
+            {
+                return VungleConsent.DENIED;
+            }
+            return VungleConsent.UNKNOWN;
         }
     }
 }